Moving modal components separately into subdir (task #6034)
