+++ resources/src/components/modals/EventView.vue
